Removed unneeded cast
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@27378 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 252c44beb6f84216024f9ebdaca453b9f63a3a0d

+++ server/fragmentation/base/src/de/uniol/inf/is/odysseus/server/fragmentation/horizontal/physicaloperator/AbstractStaticFragmentPO.java
@@ -65,7 +65,7 @@ public abstract class AbstractStaticFragmentPO<T extends IStreamObject<IMetaAttr
     
     public AbstractStaticFragmentPO(int numFragments,int heartbeatRate) {
     	super();
-        this.numFragments = (int) numFragments;
+        this.numFragments = numFragments;
         this.heartbeatCounter = new int[this.numFragments];
         this.heartbeatRate = heartbeatRate;
         this.elementsSend = new long[this.numFragments];
